+++ attributes/default.rb
@@ -6,6 +6,11 @@ default['java']['oracle']['accept_oracle_download_terms'] = true
 default['pingfed']['admin_port'] = 9999
 default['pingfed']['listen_port'] = 9031
 default['pingfed']['bind_address'] = '0.0.0.0'
+default['pingfed']['default_cluster_bind_port'] = 7600
+default['pingfed']['console_cluster_bind_port'] = 7610
+default['pingfed']['cluster_failure_detection_bind_port'] = 7700
+default['pingfed']['cluster_password'] = '123456789012345678901234567890'
+
 default['pingfed']['operational_mode']['standalone'] = 'STANDALONE'
 
 default['pingfed']['install_dir'] = '/usr/local'