+++ metadata.rb
@@ -3,7 +3,7 @@ maintainer 'Scott Marshall'
 maintainer_email 'scott.marshall@johnmuirhealth.com'
 description 'Installs/Configures pingfederate'
 long_description 'Installs/Configures pingfederate'
-version '0.2.0'
+version '0.3.0'
 
 depends 'java'
 depends 'systemd'