+++ recipes/engine_instance.rb
@@ -1,0 +1,9 @@
+include_recipe 'pingfederate::prerequisites'
+include_recipe 'pingfederate::cluster_prerequisites'
+
+service 'pingfederate-engine'
+
+pingfederate_install "engine" do
+  server_type "engine"
+end
+