+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # 0.2.0
 - upgrade pingfed 9.2.2
 - change long wait to wait that port is open
+- fixed restart issue with systemd
 
 # 0.1.3
 - Added sleep 2 minutes at first install for API calls to respond